+++ index.js
@@ -4,6 +4,6 @@
 import 'react-native-gesture-handler';
 import {AppRegistry} from 'react-native';
 import {name as appName} from './app.json';
-import App from './src/screens/MyCar/App';
+import App from './src/screens/CarDetails/App';
 
 AppRegistry.registerComponent(appName, () => App);
