+++ src/locals/he.json
@@ -19,7 +19,5 @@
   "convenienceInfo":"לנוחיותך איגדנו מספר מדריכים לתפעול הרכב",
   "instructions":"מדריך",
   "documentSent":"מסמכים נשלחו",
-  "documtSentInfo": "המסמכים נשלחו בהצלחה למייל OirS@gmail.com",
-  "shareMyDocuments":"שיתוף המסמכים שלי",
-  "chooseFilesText":"בחרו את המסמכים אותם תרצו לשתף"
+  "documtSentInfo": "המסמכים נשלחו בהצלחה למייל OirS@gmail.com"
 }
