+++ src/screens/CarDetails/components/AirPressureGuide.js
@@ -84,3 +84,5 @@ const styles = StyleSheet.create({
         fontWeight: 'bold',
       },
 });
+
+export default AirPressureGuide;
