+++ src/screens/CarDetails/components/FirstCarDetails.js
@@ -132,3 +132,5 @@ const styles = StyleSheet.create({
     marginVertical: 10,
   },
 });
+
+export default FirstCarDetails;
