+++ src/screens/CarDetails/components/SecondCarDetails.js
@@ -100,3 +100,5 @@ const styles = StyleSheet.create({
     color:'white',
   }
 });
+
+export default SecondCarDetails;
