+++ src/screens/CarDetails/components/Service.js
@@ -73,3 +73,5 @@ const styles = StyleSheet.create({
     fontSize:12
   }
 });
+
+export default Service;
